move general to component
