+++ frontend/src/components/ui/copy-field.ts
@@ -36,6 +36,9 @@ export class CopyField extends TailwindElement {
   hoist = false;
 
   @property({ type: Boolean })
+  border = true;
+
+  @property({ type: Boolean })
   monostyle = true;
 
   @property({ type: Boolean })
@@ -65,7 +68,7 @@ export class CopyField extends TailwindElement {
       <div
         role="group"
         class=${clsx(
-          tw`rounded border`,
+          this.border && tw`rounded border`,
           this.filled && tw`bg-slate-50`,
           this.monostyle && tw`font-monostyle`,
         )}
